+++ shardingsphere-infra/shardingsphere-infra-common/src/main/resources/META-INF/services/org.apache.shardingsphere.infra.metadata.model.physical.jdbc.handler.DatabaseSpecialHandler
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.infra.metadata.model.physical.jdbc.handler.impl.OracleTableNamePatternHandler
+org.apache.shardingsphere.infra.metadata.model.physical.jdbc.handler.impl.OracleDatabaseSpecialHandler
